+++ changelog.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `Eq` and `Ord` instances for `Operator`
+- `Convex.BuildTx`: Added the option to look at all of a transaction's inputs when building a transaction
 
 ### Deleted